+++ utils.go
@@ -124,4 +124,3 @@ func convertValue(value interface{}) driver.Value {
 		return fmt.Sprintf("%v", v)
 	}
 }
-